Ensure deletes and inserts happen in the correct order

+++ config/worship-services-sensitive-consumer/positions-dispatching/delta-sync-dispatching.js
@@ -31,39 +31,36 @@ async function dispatch(lib, data) {
     const insertStatements = inserts.map(o => `${o.subject} ${o.predicate} ${o.object}.`);
 
     if (deleteStatements.length) {
-      await transformStatements(fetch, deleteStatements).then(
-        transformedStatements => {
-          deleteFromAllGraphs(
-            muAuthSudo.updateSudo,
-            transformedStatements,
-            { 'mu-call-scope-id': 'http://redpencil.data.gift/id/concept/muScope/deltas/write-for-dispatch' },
-            process.env.MU_SPARQL_ENDPOINT, //Note: this is the default endpoint through auth
-            MAX_DB_RETRY_ATTEMPTS,
-            SLEEP_BETWEEN_BATCHES,
-            SLEEP_TIME_AFTER_FAILED_DB_OPERATION,
-          );
-        }
-      )
+      const transformedStatementsToDelete = await transformStatements(fetch, deleteStatements);
+      if (transformedStatementsToDelete.length) {
+        await deleteFromAllGraphs(
+          muAuthSudo.updateSudo,
+          transformedStatementsToDelete,
+          { 'mu-call-scope-id': 'http://redpencil.data.gift/id/concept/muScope/deltas/write-for-dispatch' },
+          process.env.MU_SPARQL_ENDPOINT, //Note: this is the default endpoint through auth
+          MAX_DB_RETRY_ATTEMPTS,
+          SLEEP_BETWEEN_BATCHES,
+          SLEEP_TIME_AFTER_FAILED_DB_OPERATION,
+        );
+      }
     }
 
-
     if (insertStatements.length) {
-      await transformStatements(fetch, insertStatements).then(
-        transformedStatements => {
-          batchedDbUpdate(
-            muAuthSudo.updateSudo,
-            INGEST_GRAPH,
-            transformedStatements,
-            { 'mu-call-scope-id': 'http://redpencil.data.gift/id/concept/muScope/deltas/write-for-dispatch' },
-            process.env.MU_SPARQL_ENDPOINT, //Note: this is the default endpoint through auth
-            BATCH_SIZE,
-            MAX_DB_RETRY_ATTEMPTS,
-            SLEEP_BETWEEN_BATCHES,
-            SLEEP_TIME_AFTER_FAILED_DB_OPERATION,
-            "INSERT"
-          );
-        }
-      )
+      const transformedStatementsToInsert = await transformStatements(fetch, insertStatements);
+      if (transformedStatementsToInsert.length) {
+        await batchedDbUpdate(
+          muAuthSudo.updateSudo,
+          INGEST_GRAPH,
+          transformedStatementsToInsert,
+          { 'mu-call-scope-id': 'http://redpencil.data.gift/id/concept/muScope/deltas/write-for-dispatch' },
+          process.env.MU_SPARQL_ENDPOINT, //Note: this is the default endpoint through auth
+          BATCH_SIZE,
+          MAX_DB_RETRY_ATTEMPTS,
+          SLEEP_BETWEEN_BATCHES,
+          SLEEP_TIME_AFTER_FAILED_DB_OPERATION,
+          "INSERT"
+        );
+      }
     }
   }
 }
